+++ tengine/game/camera.cpp
@@ -35,8 +35,8 @@ void CCamera::Think()
 		Vector vecForward, vecRight;
 		AngleVectors(m_angFreeCamera, &vecForward, NULL, &vecRight);
 
-		m_vecFreeCamera += vecForward * m_vecFreeVelocity.x * GameServer()->GetFrameTime() * 20;
-		m_vecFreeCamera += vecRight * m_vecFreeVelocity.z * GameServer()->GetFrameTime() * 20;
+		m_vecFreeCamera += vecForward * m_vecFreeVelocity.x * (float)GameServer()->GetFrameTime() * 20;
+		m_vecFreeCamera += vecRight * m_vecFreeVelocity.z * (float)GameServer()->GetFrameTime() * 20;
 	}
 	else
 	{
